Move DefinitionLinesDenylistEnforcer to a separate package
