+++ main/pipeline/definition_checker/BUILD
@@ -1,0 +1,12 @@
+cc_library(
+    name = "definition_checker",
+    srcs = [
+        "DefinitionLinesDenylistEnforcer.cc",
+    ],
+    hdrs = ["DefinitionLinesDenylistEnforcer.h"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//ast/treemap",
+        "//core",
+    ],
+)
